Rename ledger_live connector
ledger_live -> ledgerLive
