+++ src/web3/connectors/index.ts
@@ -2,6 +2,6 @@ export * from "./taho"
 export * from "./metamask"
 export * from "./walletConnect"
 export * from "./coinbaseWallet"
-export * from "./ledger_live"
+export * from "./ledgerLive"
 export { AbstractConnector } from "@web3-react/abstract-connector"
 export { UserRejectedRequestError } from "@web3-react/injected-connector"
